+++ packages/icons/src/SvgIcon/types.ts
@@ -6,9 +6,17 @@ export type SVGProp = Omit<HTMLProps<HTMLOrSVGElement>, 'size'>;
 export interface SvgIconProps extends SVGProp, WithThemeProp {
     /** Icon Size */
     size?: IconSizes;
+    /** Icon variants */
+    variant?: 'solid' | 'flat';
     /** Icon color */
-    color?: string;
-    /** Show hover state if you want to use it as individual button */
+    iconColor?: string;
+    /** Background color */
+    bgColor?: string;
+    /** Background color when we hover on it */
+    hoverBgColor?: string;
+    /** Icon color when we hover on it */
+    hoverIconColor?: string;
+    /** Set it true to show the hover effect */
     withHoverEffect?: boolean;
     /** Margin for the Icon */
     margin?: string;